Add put and delete routes to API. Remove old server code
