--- server/src/config/default.json
@@ -1,11 +1,0 @@
-{
-    "appDash": {
-        "port": 9090,
-        "authorize": false,
-        "adminPass": "someSecret",
-        "jwtSecret": "someSecret"
-    },
-    "mongo": {
-        "connection": ""
-    }
-}
